mmdbctl structured like ipinfo/cli
