+++ cmd_import.go
@@ -1,22 +1,11 @@
 package main
 
 import (
-	"bufio"
-	"encoding/csv"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	"net"
-	"os"
-	"path/filepath"
-	"strings"
 
 	"github.com/ipinfo/cli/lib/complete"
 	"github.com/ipinfo/cli/lib/complete/predict"
-	"github.com/maxmind/mmdbwriter"
-	"github.com/maxmind/mmdbwriter/inserter"
-	"github.com/maxmind/mmdbwriter/mmdbtype"
+	"github.com/ipinfo/mmdbctl/lib"
 	"github.com/spf13/pflag"
 )
 
@@ -154,417 +143,9 @@ Options:
 }
 
 func cmdImport() error {
-	var fIn string
-	var fOut string
-	var fCsv bool
-	var fTsv bool
-	var fJson bool
-	var fFields []string
-	var fFieldsFromHdr bool
-	var fRangeMultiCol bool
-	var fJoinKeyCol bool
-	var fNoFields bool
-	var fNoNetwork bool
-	var fIp int
-	var fSize int
-	var fMerge string
-	var fIgnoreEmptyVals bool
-	var fDisallowReserved bool
-	var fAlias6to4 bool
-	var fDisableMetadataPtrs bool
-
-	_h := "see description in --help"
-	pflag.StringVarP(&fIn, "in", "i", "", _h)
-	pflag.StringVarP(&fOut, "out", "o", "", _h)
-	pflag.BoolVarP(&fCsv, "csv", "c", false, _h)
-	pflag.BoolVarP(&fTsv, "tsv", "t", false, _h)
-	pflag.BoolVarP(&fJson, "json", "j", false, _h)
-	pflag.StringSliceVarP(&fFields, "fields", "f", nil, _h)
-	pflag.BoolVar(&fFieldsFromHdr, "fields-from-header", false, _h)
-	pflag.BoolVar(&fRangeMultiCol, "range-multicol", false, _h)
-	pflag.BoolVar(&fJoinKeyCol, "joinkey-col", false, _h)
-	pflag.BoolVar(&fNoFields, "no-fields", false, _h)
-	pflag.BoolVar(&fNoNetwork, "no-network", false, _h)
-	pflag.IntVar(&fIp, "ip", 6, _h)
-	pflag.IntVarP(&fSize, "size", "s", 32, _h)
-	pflag.StringVarP(&fMerge, "merge", "m", "none", _h)
-	pflag.BoolVar(&fIgnoreEmptyVals, "ignore-empty-values", false, _h)
-	pflag.BoolVar(&fDisallowReserved, "disallow-reserved", false, _h)
-	pflag.BoolVar(&fAlias6to4, "alias-6to4", false, _h)
-	pflag.BoolVar(&fDisableMetadataPtrs, "disable-metadata-pointers", true, _h)
+	f := lib.CmdImportFlags{}
+	f.Init()
 	pflag.Parse()
 
-	// help?
-	if fHelp || (pflag.NArg() == 1 && pflag.NFlag() == 0) {
-		printHelpImport()
-		return nil
-	}
-
-	// optional input as 1st and output as 2nd argument.
-	if pflag.NArg() >= 2 {
-		fIn = pflag.Arg(1)
-		fOut = pflag.Arg(2)
-	}
-
-	// validate IP version.
-	if fIp != 4 && fIp != 6 {
-		return errors.New("ip version must be \"4\" or \"6\"")
-	}
-
-	// validate record size.
-	if fSize != 24 && fSize != 28 && fSize != 32 {
-		return errors.New("record size must be 24, 28 or 32")
-	}
-
-	// validate merge strategy.
-	var mergeStrategy inserter.FuncGenerator
-	if fMerge == "none" {
-		mergeStrategy = inserter.ReplaceWith
-	} else if fMerge == "toplevel" {
-		mergeStrategy = inserter.TopLevelMergeWith
-	} else if fMerge == "recurse" {
-		mergeStrategy = inserter.DeepMergeWith
-	} else {
-		return errors.New("merge strategy must be \"none\", \"toplevel\" or \"recurse\"")
-	}
-
-	// figure out file type.
-	var delim rune
-	if !fCsv && !fTsv && !fJson {
-		if strings.HasSuffix(fIn, ".csv") {
-			delim = ','
-		} else if strings.HasSuffix(fIn, ".tsv") {
-			delim = '\t'
-		} else if strings.HasSuffix(fIn, ".json") {
-			delim = '-'
-		} else {
-			return errors.New("input file type unknown")
-		}
-	} else {
-		if fCsv && fTsv || fCsv && fJson || fTsv && fJson {
-			return errors.New("multiple input file types specified")
-		} else if fCsv {
-			delim = ','
-		} else if fTsv {
-			delim = '\t'
-		} else {
-			delim = '-'
-		}
-	}
-
-	// figure out fields.
-	fieldSrcCnt := 0
-	if fFields != nil && len(fFields) > 0 {
-		fieldSrcCnt += 1
-	}
-	if fFieldsFromHdr {
-		fieldSrcCnt += 1
-	}
-	if fNoFields {
-		fieldSrcCnt += 1
-	}
-	if fieldSrcCnt > 1 {
-		return errors.New("conflicting field sources specified.")
-	}
-	if fNoFields {
-		fFields = []string{}
-		fNoNetwork = false
-	} else if !fFieldsFromHdr && (fFields == nil || len(fFields) == 0) {
-		fFieldsFromHdr = true
-	}
-
-	if fJoinKeyCol {
-		fRangeMultiCol = true
-	}
-
-	// prepare output file.
-	var outFile *os.File
-	if fOut == "" {
-		outFile = os.Stdout
-	} else {
-		var err error
-		outFile, err = os.Create(fOut)
-		if err != nil {
-			return fmt.Errorf("could not create %v: %w", fOut, err)
-		}
-		defer outFile.Close()
-	}
-
-	// init tree.
-	dbdesc := "ipinfo " + filepath.Base(fOut)
-	tree, err := mmdbwriter.New(
-		mmdbwriter.Options{
-			DatabaseType: dbdesc,
-			Description: map[string]string{
-				"en": dbdesc,
-			},
-			Languages:               []string{"en"},
-			DisableIPv4Aliasing:     !fAlias6to4,
-			IncludeReservedNetworks: !fDisallowReserved,
-			IPVersion:               fIp,
-			RecordSize:              fSize,
-			DisableMetadataPointers: fDisableMetadataPtrs,
-			Inserter:                mergeStrategy,
-		},
-	)
-	if err != nil {
-		return fmt.Errorf("could not create tree: %w", err)
-	}
-
-	// prepare input file.
-	var inFile *os.File
-	if fIn == "" || fIn == "-" {
-		inFile = os.Stdin
-	} else {
-		var err error
-		inFile, err = os.Open(fIn)
-		if err != nil {
-			return fmt.Errorf("invalid input file %v: %w", fIn, err)
-		}
-		defer inFile.Close()
-	}
-
-	inFileBuffered := bufio.NewReaderSize(inFile, 65536)
-
-	entrycnt := 0
-	if delim == ',' || delim == '\t' {
-		var rdr reader
-		if delim == ',' {
-			csvrdr := csv.NewReader(inFileBuffered)
-			csvrdr.Comma = delim
-			csvrdr.LazyQuotes = true
-
-			rdr = csvrdr
-		} else {
-			tsvrdr := NewTsvReader(inFileBuffered)
-
-			rdr = tsvrdr
-		}
-
-		// read from input, scanning & parsing each line according to delim,
-		// then insert that into the tree.
-		dataColStart := 1
-		hdrSeen := false
-		for {
-			parts, err := rdr.Read()
-			if err == io.EOF {
-				break
-			} else if err != nil {
-				return fmt.Errorf("input scanning failed: %w", err)
-			}
-
-			// on header line?
-			if !hdrSeen {
-				hdrSeen = true
-
-				// check if the header has a multi-column range.
-				if len(parts) > 1 && parts[0] == "start_ip" && parts[1] == "end_ip" {
-					fRangeMultiCol = true
-
-					// maybe we also have a join key?
-					if len(parts) > 2 && parts[2] == "join_key" {
-						fJoinKeyCol = true
-					}
-				}
-
-				if fRangeMultiCol {
-					if fJoinKeyCol {
-						dataColStart = 3
-					} else {
-						dataColStart = 2
-					}
-				}
-
-				// need to get fields from hdr?
-				if fFieldsFromHdr {
-					// skip all non-data columns.
-					fFields = parts[dataColStart:]
-				}
-
-				// insert empty values for all fields in 0.0.0.0/0 if requested.
-				if fIgnoreEmptyVals {
-					_, network, _ := net.ParseCIDR("0.0.0.0/0")
-					record := mmdbtype.Map{}
-					for _, field := range fFields {
-						record[mmdbtype.String(field)] = mmdbtype.String("")
-					}
-					if err := tree.Insert(network, record); err != nil {
-						return errors.New(
-							"couldn't insert empty values to 0.0.0.0/0",
-						)
-					}
-				}
-
-				// should we skip this first line now?
-				if fFieldsFromHdr {
-					continue
-				}
-			}
-
-			networkStr := parts[0]
-
-			// convert 2 IPs into IP range?
-			if fRangeMultiCol {
-				networkStr = parts[0] + "-" + parts[1]
-			}
-
-			// add network part to single-IP network if it's missing.
-			isNetworkRange := strings.Contains(networkStr, "-")
-			if !isNetworkRange && !strings.Contains(networkStr, "/") {
-				if fIp == 6 && strings.Contains(networkStr, ":") {
-					networkStr += "/128"
-				} else {
-					networkStr += "/32"
-				}
-			}
-
-			// prep record.
-			record := mmdbtype.Map{}
-			if !fNoNetwork {
-				record["network"] = mmdbtype.String(networkStr)
-			}
-			for i, field := range fFields {
-				record[mmdbtype.String(field)] = mmdbtype.String(parts[i+dataColStart])
-			}
-
-			// range insertion or cidr insertion?
-			if isNetworkRange {
-				networkStrParts := strings.Split(networkStr, "-")
-				startIp := net.ParseIP(networkStrParts[0])
-				endIp := net.ParseIP(networkStrParts[1])
-				if err := tree.InsertRange(startIp, endIp, record); err != nil {
-					fmt.Fprintf(
-						os.Stderr, "warn: couldn't insert line '%v'\n",
-						strings.Join(parts, string(delim)),
-					)
-				}
-			} else {
-				_, network, err := net.ParseCIDR(networkStr)
-				if err != nil {
-					return fmt.Errorf(
-						"couldn't parse cidr \"%v\": %w",
-						networkStr, err,
-					)
-				}
-				if err := tree.Insert(network, record); err != nil {
-					fmt.Fprintf(
-						os.Stderr, "warn: couldn't insert line '%v'\n",
-						strings.Join(parts, string(delim)),
-					)
-				}
-			}
-
-			entrycnt += 1
-		}
-	} else if delim == '-' {
-		dataStream := json.NewDecoder(inFileBuffered)
-		for {
-			// Decode one JSON document.
-			var row interface{}
-			err := dataStream.Decode(&row)
-
-			if err != nil {
-				// io.EOF is expected at end of stream.
-				if err != io.EOF {
-					return fmt.Errorf("error in io.EOF: %w", err)
-				}
-				break
-			}
-			mResult := row.(map[string]interface{})
-
-			// insert empty values for all fields in 0.0.0.0/0 if requested.
-			if fIgnoreEmptyVals {
-				_, network, _ := net.ParseCIDR("0.0.0.0/0")
-				record := mmdbtype.Map{}
-				for _, field := range fFields {
-					record[mmdbtype.String(field)] = mmdbtype.String("")
-				}
-				if err := tree.Insert(network, record); err != nil {
-					return errors.New(
-						"couldn't insert empty values to 0.0.0.0/0",
-					)
-				}
-			}
-
-			// convert 2 IPs into IP range?
-			var networkStr string
-			if val, ok := mResult["start_ip"].(string); ok {
-				networkStr = val + "-" + mResult["end_ip"].(string)
-				delete(mResult, "start_ip")
-				delete(mResult, "end_ip")
-				if _, ok := mResult["join_key"].(string); ok {
-					delete(mResult, "join_key")
-				}
-			} else if val, ok := mResult["range"].(string); ok {
-				networkStr = val
-				delete(mResult, "range")
-			} else {
-				return errors.New(
-					"couldn't get ip or range from the record",
-				)
-			}
-
-			// add network part to single-IP network if it's missing.
-			isNetworkRange := strings.Contains(networkStr, "-")
-			if !isNetworkRange && !strings.Contains(networkStr, "/") {
-				if fIp == 6 && strings.Contains(networkStr, ":") {
-					networkStr += "/128"
-				} else {
-					networkStr += "/32"
-				}
-			}
-
-			// prep record.
-			record := mmdbtype.Map{}
-			if !fNoNetwork {
-				record["network"] = mmdbtype.String(networkStr)
-			}
-
-			mResultStr := mapInterfaceToStr(mResult)
-			for k, v := range mResultStr {
-				record[mmdbtype.String(k)] = mmdbtype.String(v)
-			}
-
-			// range insertion or cidr insertion?
-			if isNetworkRange {
-				networkStrParts := strings.Split(networkStr, "-")
-				startIp := net.ParseIP(networkStrParts[0])
-				endIp := net.ParseIP(networkStrParts[1])
-				if err := tree.InsertRange(startIp, endIp, record); err != nil {
-					fmt.Fprintf(
-						os.Stderr, "warn: couldn't insert '%v'\n",
-						mResult,
-					)
-				}
-			} else {
-				_, network, err := net.ParseCIDR(networkStr)
-				if err != nil {
-					return fmt.Errorf(
-						"couldn't parse cidr \"%v\": %w",
-						networkStr, err,
-					)
-				}
-				if err := tree.Insert(network, record); err != nil {
-					fmt.Fprintf(
-						os.Stderr, "warn: couldn't insert '%v'\n",
-						mResult,
-					)
-				}
-			}
-
-			entrycnt += 1
-		}
-	}
-
-	if entrycnt == 0 {
-		return errors.New("nothing to import")
-	}
-
-	// write out mmdb file.
-	fmt.Fprintf(os.Stderr, "writing to %s (%v entries)\n", fOut, entrycnt)
-	if _, err := tree.WriteTo(outFile); err != nil {
-		return fmt.Errorf("writing out to tree failed: %w", err)
-	}
-
-	return nil
+	return lib.CmdImport(f, pflag.Args()[1:], printHelpImport)
 }
