+++ lib/reader.go
@@ -1,4 +1,4 @@
-package main
+package lib
 
 type reader interface {
 	Read() (record []string, err error)
