+++ lib/writer.go
@@ -1,4 +1,4 @@
-package main
+package lib
 
 type writer interface {
 	Write(record []string) error
